refactor(admin): split quantercise overview panels
Move the Quantercise dashboard display panels out of the overview route. The page now owns environment lookup, dashboard fetching, suspense fallback, and error handling while metrics, alerts, quick stats, activity, content pipeline, and last-updated panels live in a focused module.

Verified: pnpm exec prettier --check apps/admin/src/app/quantercise/page.tsx apps/admin/src/app/quantercise/dashboard-panels.tsx

Verified: git diff --check -- apps/admin/src/app/quantercise/page.tsx apps/admin/src/app/quantercise/dashboard-panels.tsx

Verified: pnpm --filter @anipotts/admin typecheck

Verified: pnpm --filter @anipotts/admin lint

Verified: pnpm --filter @anipotts/admin build

Verified: pnpm validate

Verified: pnpm dlx knip --no-exit-code
